improve: split schemas.py into domain modules
- api/schemas.py → re-export facade for backward compatibility
- api/schemas_core.py: sessions, events, checkpoints, traces, API keys
- api/schemas_alerts.py: anomaly alerts, alert policies, fix notes
- api/schemas_analysis.py: workflow, safety, redundancy, causal, drift, baseline
